requirements.txt
asgiref==3.3.1
autopep8==1.5.7
Django==3.1.6
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-rest-multiple-models==2.1.3
django-summernote==0.8.11.6
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
